test3a works ~
test3a works however the image in t.right(90) off

+++ Project_03/lib/point.c
@@ -53,8 +53,8 @@ void point_draw(Point *p, Image *src, Color c){
         printf("ERROR: point_draw >> pointer parameter is NULL\n");
         exit(-1);
     }
-    double x=p->val[0];
-    double y=p->val[1];
+    int x=(int)p->val[0];
+    int y=(int)p->val[1];
     image_setc(src, x, y, 0, c.rgb[0]);
     image_setc(src, x, y, 1, c.rgb[1]);
     image_setc(src, x, y, 2, c.rgb[2]);
